+++ app/src/main/java/com/adam/leoadapterapp/MainActivity.kt
@@ -21,6 +21,3 @@ class MainActivity : ComponentActivity() {
         }
     }
 }
-
-
-