+++ library/core/src/main/kotlin/com/adkhambek/leo/LeoAdapterDSL.kt
@@ -1,0 +1,7 @@
+package com.adkhambek.leo
+
+import androidx.annotation.RestrictTo
+
+@DslMarker
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+public annotation class LeoAdapterDSL