+++ library/leo/src/main/kotlin/com/adkhambek/leo/LeoAdapter.kt
@@ -1,0 +1,10 @@
+package com.adkhambek.leo
+
+import androidx.recyclerview.widget.RecyclerView
+import androidx.viewbinding.ViewBinding
+
+public abstract class LeoAdapter<T : Any, VB : ViewBinding> : RecyclerView.Adapter<LeoVH<VB>>() {
+
+    @LeoAdapterDSL
+    public abstract fun setList(data: List<T>)
+}