+++ library/paging/src/main/kotlin/com/adkhambek/leo/LeoPagingAdapter.kt
@@ -1,0 +1,9 @@
+package com.adkhambek.leo
+
+import androidx.paging.PagingDataAdapter
+import androidx.recyclerview.widget.DiffUtil
+import androidx.viewbinding.ViewBinding
+
+public abstract class LeoPagingAdapter<T : Any, VB : ViewBinding>(
+    diffCallback: DiffUtil.ItemCallback<T>,
+) : PagingDataAdapter<T, LeoVH<VB>>(diffCallback)